refactor: extract format_cef escape helpers (#540)
format_cef.go was 809 lines — over the project's ~300-line
guideline. Move the string-returning escape helpers and the
extension-key validator into a new format_cef_escape.go file. Pure
refactor; no behaviour change, no public API change.

Kept in format_cef.go (725 lines):
- cefBufPool sync.Pool, putCEFBuf
- DefaultCEFFieldMapping helpers
- CEFFormatter struct + all methods (Format, formatBuf,
  writeFieldExtensions, severity, description, fieldMapping,
  mergedMapping, SetFrameworkFields, writeFrameworkExtensions)
- noCopy type
- writeExtField family (writeExtField, writeExtFieldInt64,
  writeExtFieldValue)
- writeEscapedExtValueString (in-place escape, tightly coupled
  to appendFormatFieldValue)
- appendFormatFieldValue, mapFieldKey

Moved to format_cef_escape.go (114 lines):
- cefEscapeHeader (escapes \, |, \n, \r in header fields)
- cefEscapeExtValue (escapes \, =, control chars in ext values)
- validateExtKey (one-shot extension-key validator)

format_cef.go remains over the 300-line target — accepted per the
same rationale as audit.go and syslog.go (#540): the CEFFormatter
methods and writeExt* family are tightly coupled to
appendFormatFieldValue and the in-place escape variant; further
splitting would harm cohesion of the format hot path.

Verification:
- go build ./... → clean
- go test -race -count=1 . → green (180s)

Refs #540.
